+++ OneSignalExample/Assets/OneSignal/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added value to additionalData on Android for NotificationOpened
 - Reverted [#430](https://github.com/OneSignal/OneSignal-Unity-SDK/pull/430) due to a deprecation of where Android resources can be stored in Unity. Notification icons to be changed for Android can again be found at `Assets/Plugins/Android/OneSignalConfig.plugin`.
 - Updated example code for PostNotification to show an example that works without the API key
+- Reimplemented support for `RemoveExternalUserId`
 
 ## [3.0.0]
 ### Changed